pyproject.toml
[project]
name = "Speedrun-Auto-Upload"
version = "0.1.0"
description = "Automatically find the record run, extract the recording and upload the run"
authors = [
    {name = "BrainStone", email = "yannick@tedworld.de"},
]
dependencies = [
    # "livesplit-parser>=0.2.1",
    "livesplit-parser @ git+https://github.com/BrainStone/livesplit_parser.git@custom_release",
    "ffmpeg-python>=0.2.0",
    "dask>=2024.4.1",
    "dask-expr>=1.0.10",
    "tzlocal>=5.2",
    "google-auth>=2.29.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.131.0",
    "tqdm>=4.66.4",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "GPLv3"}


[tool.pdm]
distribution = false
